provaSem5Luiz/src/requirements.txt
tinydb-4.8.0
flask-3.0.2
